projects/who-am-ai/requirements.txt
python-dotenv
openai
pydantic
pypdf
gradio
